2014-02-23 15:43 UTC+0100 Viktor Szakáts (vszakats users.noreply.github.com)   * config/darwin/clang.mk   * utils/hbmk2/hbmk2.prg     + darwin/clang warning options upped from -Wall to -Weverything       plus an explicit list of exclusions (can be extended/reduced over time)       This will result in extra warnings.
  * src/compiler/complex.c
    * fixed wrong cast

  * src/rtl/cdpapi.c
  * src/rtl/gtchrmap.c
  * src/rtl/hbbfish.c
  * src/rtl/hbgtcore.c
  * src/vm/garbage.c
    ! added missing static keywords to local data

  * src/rtl/hbjson.c
    ! fixed variable redeclared in sub-block

  * contrib/hbmisc/spd.c
    * whitespace

+++ src/compiler/complex.c
@@ -445,7 +445,7 @@ static int hb_comp_dayTimeDecode( PHB_COMP_LEX pLex, PHB_PP_TOKEN pToken,
 
       if( pTime )
       {
-         lDate = hb_dateEncode( ( long ) lYear, ( long ) lMonth, ( long ) lDay );
+         lDate = hb_dateEncode( ( int ) lYear, ( int ) lMonth, ( int ) lDay );
          if( lDate != 0 || ( lYear == 0 && lMonth == 0 && lDay == 0 ) )
          {
             iType = NUM_DATE;
